Refactor MarkdownPage to be reusable and accept content as a prop
