+++ app/(pages)/blogs/[...slug]/page.tsx
