+++ app/(pages)/works/page.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import type { Metadata } from "next";
 
-import Works from "../section/work/Works";
+import Works from "../../section/work/Works";
 
 export const metadata: Metadata = {
   title: "Works | Suparna Adhikari | Full-Stack Developer & Software Engineer",
